cortex-cli/requirements.txt
# Kern-Abhängigkeiten
click>=8.0.0
pyyaml>=6.0.0
rich>=12.0.0
requests>=2.28.0

# Optionale Abhängigkeiten für volles Feature-Set
aiohttp>=3.8.0
watchdog>=2.1.0
aiofiles>=0.8.0
asyncio-mqtt>=0.11.0
python-dateutil>=2.8.0
schedule>=1.2.0

# Test-Abhängigkeiten (nur für Entwicklung)
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
